supprimer que si iduser == id lastmoodification

+++ src/Controller/GamesController.php
@@ -78,9 +78,19 @@ class  GamesController extends AbstractController{
         return $this->redirectToRoute('app_games_show_all');
     }
     #[Route('/games/delete/{id}', name: 'app_games_delete')]
-    public function delete(EntityManagerInterface $entityManager,GamesInfo $game){
-        $entityManager->remove($game);
-        $entityManager->flush();
-        return $this->redirectToRoute('app_games_show_all');
+    public function delete(EntityManagerInterface $entityManager,GamesInfo $game,Security $security){
+    $iduser=$game->getLastModificationUser();
+
+        $useractual = $security->getUser();
+    $iduserlogin=$useractual->getID();
+
+        if($iduserlogin==$iduser){
+            $entityManager->remove($game);
+            $entityManager->flush();
+            return $this->redirectToRoute('app_games_show_all');
+        }else{
+            return $this->redirectToRoute('app_games_show_all');
+        }
+
     }
 }
